# contracts/feedback/openapi.yaml
openapi: 3.1.0
...
info:
  title: Feedback Service API
  version: 1.0.0-alpha.1
  description: |
    ホスト業務システムから独立したフィードバック基盤の公開契約。
    Web GIS の project、GIS、業務 CRUD、タイル API を含めない。
servers:
  - url: /feedback/v1
security:
  - bearerAuth: []
...
paths:
...
  /sessions:
    get:
      tags: [sessions]
      operationId: listFeedbackSessions
      parameters:
        - $ref: "#/components/parameters/ApplicationKeyQuery"
        - $ref: "#/components/parameters/EnvironmentKeyQuery"
        - $ref: "#/components/parameters/ExternalWorkspaceKeyQuery"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/FeedbackSessionStatus"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: セッション一覧
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackSessionPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [sessions]
      operationId: createFeedbackSession
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackSessionCreateRequest"
      responses:
        "201":
          description: 作成したセッション
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackSessionV1"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /sessions/{sessionId}:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    get:
      tags: [sessions]
      operationId: getFeedbackSession
      responses:
        "200":
          description: セッション詳細
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackSessionV1"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [sessions]
      operationId: patchFeedbackSession
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: "#/components/schemas/FeedbackSessionPatchRequest"
      responses:
        "200":
          description: 更新したセッション
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackSessionV1"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
...
  /retention-policy:
    get:
      tags: [administration]
      operationId: getFeedbackRetentionPolicy
      parameters:
        - $ref: "#/components/parameters/ApplicationKeyQuery"
        - $ref: "#/components/parameters/ExternalWorkspaceKeyQuery"
      responses:
        "200":
          description: 保存方針
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackRetentionPolicy"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    patch:
      tags: [administration]
      operationId: patchFeedbackRetentionPolicy
      parameters:
        - $ref: "#/components/parameters/ApplicationKeyQuery"
        - $ref: "#/components/parameters/ExternalWorkspaceKeyQuery"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: "#/components/schemas/FeedbackRetentionPolicy"
      responses:
        "200":
          description: 更新した保存方針
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackRetentionPolicy"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        直接OIDC JWTまたは契約済みtoken exchange JWT。直接OIDC JWTは検証対象のissuer、audience、
        時刻、設定済みsubject claimに加え、`feedback_permissions`を1件以上のFeedbackPermissionからなる
        文字列配列として必須とする。直接OIDCのpermissionは全resourceでDB membershipと交差する。
        token exchange JWTのresource scopeとclaimはtoken exchange公開契約に従う。
  parameters:
...
    ApplicationKeyQuery:
      name: applicationKey
      in: query
      required: true
      schema: { type: string, pattern: "^[a-z][a-z0-9-]{0,62}$" }
    EnvironmentKeyQuery:
      name: environmentKey
      in: query
      required: true
      schema: { type: string, minLength: 1, maxLength: 100 }
    ExternalWorkspaceKeyQuery:
      name: externalWorkspaceKey
      in: query
      required: true
      schema: { type: string, minLength: 1, maxLength: 200 }
    SessionId:
      name: sessionId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Cursor:
      name: cursor
      in: query
      schema: { type: string, maxLength: 2000 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 200 }
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema: { type: string, minLength: 3, maxLength: 200 }
...
  headers:
    ETag:
      description: resource version。更新時は If-Match へそのまま指定する。
      schema: { type: string }
  responses:
    BadRequest:
      description: 契約または manifest に適合しない入力
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/FeedbackProblem" }
    Unauthorized:
      description: 未認証または token 無効
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/FeedbackProblem" }
    Forbidden:
      description: feedback permission 不足
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/FeedbackProblem" }
    NotFound:
      description: resource 不在または存在秘匿
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/FeedbackProblem" }
    Conflict:
      description: open session 競合などの一意制約違反
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/FeedbackProblem" }
    PreconditionFailed:
      description: ETag が古い
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/FeedbackProblem" }
...
  schemas:
...
    FeedbackProblem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, requestId]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        code: { type: string, pattern: "^[a-z0-9_.-]+$" }
        requestId: { type: string, minLength: 1, maxLength: 200 }
...
    FeedbackSessionStatus:
      type: string
      enum: [draft, open, closed]
    FeedbackSessionV1:
      type: object
      additionalProperties: false
      required:
        [id, applicationKey, environmentKey, externalWorkspaceKey, manifestVersion, title, status,
         outOfScopePosting, scopes, perspectives, createdAt, updatedAt, version]
      properties:
        id: { type: string, format: uuid }
        applicationKey: { type: string }
        environmentKey: { type: string }
        externalWorkspaceKey: { type: string }
        manifestVersion: { type: string }
        title: { type: string, minLength: 1, maxLength: 200 }
        description: { type: [string, "null"], maxLength: 5000 }
        status: { $ref: "#/components/schemas/FeedbackSessionStatus" }
        outOfScopePosting: { enum: [allow, warn, deny] }
        startAt: { type: [string, "null"], format: date-time }
        endAt: { type: [string, "null"], format: date-time }
        scopes:
          type: array
          items: { $ref: "#/components/schemas/FeedbackSessionScopeV1" }
        perspectives:
          type: array
          items: { $ref: "#/components/schemas/FeedbackSessionPerspectiveV1" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        version: { type: integer, minimum: 1 }
    FeedbackSessionScopeV1:
      type: object
      additionalProperties: false
      required: [pageKey, reviewable]
      properties:
        pageKey: { type: string, minLength: 1, maxLength: 100 }
        routeTemplate: { type: [string, "null"], maxLength: 500 }
        reviewable: { type: boolean }
        perspectiveCodes:
          type: array
          uniqueItems: true
          maxItems: 100
          description: この画面で確認する観点コード。空配列または省略時は有効な観点すべて。
          items: { type: string, minLength: 1, maxLength: 100 }
    FeedbackSessionPerspectiveV1:
      type: object
      additionalProperties: false
      required: [code, label, status]
      properties:
        code: { type: string, minLength: 1, maxLength: 100 }
        label: { type: string, minLength: 1, maxLength: 200 }
        status: { enum: [active, future, out-of-scope] }
        guidance: { type: [string, "null"], maxLength: 5000 }
    FeedbackSessionPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/FeedbackSessionV1" }
        nextCursor: { type: [string, "null"] }
        totalCount: { type: integer, minimum: 0 }
    FeedbackSessionCreateRequest:
      type: object
      additionalProperties: false
      required: [applicationKey, environmentKey, externalWorkspaceKey, manifestVersion, title, outOfScopePosting]
      properties:
        applicationKey: { type: string }
        environmentKey: { type: string }
        externalWorkspaceKey: { type: string }
        manifestVersion: { type: string }
        title: { type: string, minLength: 1, maxLength: 200 }
        description: { type: [string, "null"], maxLength: 5000 }
        status:
          allOf: [{ $ref: "#/components/schemas/FeedbackSessionStatus" }]
          default: draft
        outOfScopePosting: { enum: [allow, warn, deny], default: warn }
        startAt: { type: [string, "null"], format: date-time }
        endAt: { type: [string, "null"], format: date-time }
        scopes:
          type: array
          items: { $ref: "#/components/schemas/FeedbackSessionScopeV1" }
        perspectives:
          type: array
          items: { $ref: "#/components/schemas/FeedbackSessionPerspectiveV1" }
    FeedbackSessionPatchRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        description: { type: [string, "null"], maxLength: 5000 }
        status: { $ref: "#/components/schemas/FeedbackSessionStatus" }
        outOfScopePosting: { enum: [allow, warn, deny] }
        startAt: { type: [string, "null"], format: date-time }
        endAt: { type: [string, "null"], format: date-time }
        scopes:
          type: array
          items: { $ref: "#/components/schemas/FeedbackSessionScopeV1" }
        perspectives:
          type: array
          items: { $ref: "#/components/schemas/FeedbackSessionPerspectiveV1" }
...
    FeedbackRetentionPolicy:
      type: object
      additionalProperties: false
      required: [evidenceRetentionDays]
      properties:
        evidenceRetentionDays: { type: [integer, "null"], minimum: 1, maximum: 3650 }
        exportRetentionDays: { type: integer, minimum: 1, maximum: 365, default: 7 }
...